Refactor KafkaBridge -> KafkaCommon+KafkaBridge for webhooks
