+++ internal/kldkafka/client.go
@@ -15,31 +15,29 @@
 package kldkafka
 
 import (
+	"sync"
+
 	"github.com/Shopify/sarama"
 	cluster "github.com/bsm/sarama-cluster"
 )
 
-// We define minimal interfaces for the parts of the sarama/sarama-cluster
-// package we use, to allow for stubbing in our unit tests
-
-type kafkaFactory interface {
-	newClient(*KafkaBridge, *cluster.Config) (kafkaClient, error)
+// KafkaGoRoutines defines goroutines for processing Kafka messages from KafkaCommon
+type KafkaGoRoutines interface {
+	ConsumerMessagesLoop(consumer KafkaConsumer, producer KafkaProducer, wg *sync.WaitGroup)
+	ProducerErrorLoop(consumer KafkaConsumer, producer KafkaProducer, wg *sync.WaitGroup)
+	ProducerSuccessLoop(consumer KafkaConsumer, producer KafkaProducer, wg *sync.WaitGroup)
 }
 
-type kafkaClient interface {
-	newProducer(*KafkaBridge) (kafkaProducer, error)
-	newConsumer(*KafkaBridge) (kafkaConsumer, error)
-	Brokers() []*sarama.Broker
-}
-
-type kafkaProducer interface {
+// KafkaProducer provides the interface passed from KafkaCommon to produce messages (subset of sarama)
+type KafkaProducer interface {
 	AsyncClose()
 	Input() chan<- *sarama.ProducerMessage
 	Successes() <-chan *sarama.ProducerMessage
 	Errors() <-chan *sarama.ProducerError
 }
 
-type kafkaConsumer interface {
+// KafkaConsumer provides the interface passed from KafkaCommon to consume messages (subset of sarama-cluster)
+type KafkaConsumer interface {
 	Close() error
 	Messages() <-chan *sarama.ConsumerMessage
 	Notifications() <-chan *cluster.Notification
@@ -47,9 +45,19 @@ type kafkaConsumer interface {
 	MarkOffset(*sarama.ConsumerMessage, string)
 }
 
+type kafkaFactory interface {
+	newClient(*KafkaCommon, *cluster.Config) (kafkaClient, error)
+}
+
+type kafkaClient interface {
+	newProducer(*KafkaCommon) (KafkaProducer, error)
+	newConsumer(*KafkaCommon) (KafkaConsumer, error)
+	Brokers() []*sarama.Broker
+}
+
 type saramaKafkaFactory struct{}
 
-func (f saramaKafkaFactory) newClient(k *KafkaBridge, clientConf *cluster.Config) (c kafkaClient, err error) {
+func (f saramaKafkaFactory) newClient(k *KafkaCommon, clientConf *cluster.Config) (c kafkaClient, err error) {
 	if client, err := cluster.NewClient(k.Conf.Brokers, clientConf); err == nil {
 		c = &saramaKafkaClient{client: client}
 	}
@@ -64,10 +72,10 @@ func (c *saramaKafkaClient) Brokers() []*sarama.Broker {
 	return c.client.Brokers()
 }
 
-func (c *saramaKafkaClient) newProducer(k *KafkaBridge) (kafkaProducer, error) {
+func (c *saramaKafkaClient) newProducer(k *KafkaCommon) (KafkaProducer, error) {
 	return sarama.NewAsyncProducerFromClient(c.client.Client)
 }
 
-func (c *saramaKafkaClient) newConsumer(k *KafkaBridge) (kafkaConsumer, error) {
+func (c *saramaKafkaClient) newConsumer(k *KafkaCommon) (KafkaConsumer, error) {
 	return cluster.NewConsumerFromClient(c.client, k.Conf.ConsumerGroup, []string{k.Conf.TopicIn})
 }
