+++ internal/kldkafka/kafkabridge.go
@@ -15,21 +15,14 @@
 package kldkafka
 
 import (
-	"crypto/tls"
-	"crypto/x509"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"os"
-	"os/signal"
 	"sort"
-	"strconv"
-	"strings"
 	"sync"
 	"time"
 
 	"github.com/Shopify/sarama"
-	cluster "github.com/bsm/sarama-cluster"
 	"github.com/ethereum/go-ethereum/rpc"
 	"github.com/kaleido-io/ethconnect/internal/kldmessages"
 	"github.com/kaleido-io/ethconnect/internal/kldutils"
@@ -40,56 +33,20 @@ import (
 // KafkaBridge receives messages from Kafka and dispatches them to go-ethereum over JSON/RPC
 type KafkaBridge struct {
 	Conf struct {
-		Brokers            []string
-		ClientID           string
-		ConsumerGroup      string
-		InsecureSkipVerify bool
-		MaxInFlight        int
-		MaxTXWaitTime      int
-		TopicIn            string
-		TopicOut           string
-		RPC                struct {
+		MaxInFlight   int
+		MaxTXWaitTime int
+		RPC           struct {
 			URL string
 		}
-		SASL struct {
-			Username string
-			Password string
-		}
-		TLS struct {
-			ClientCertsFile string
-			CACertsFile     string
-			Enabled         bool
-			PrivateKeyFile  string
-		}
 	}
-	factory      kafkaFactory
+	kafka        *KafkaCommon
 	rpc          *rpc.Client
-	client       kafkaClient
-	signals      chan os.Signal
-	consumer     kafkaConsumer
-	consumerWG   sync.WaitGroup
-	producer     kafkaProducer
-	producerWG   sync.WaitGroup
-	saramaLogger saramaLogger
 	processor    MsgProcessor
 	inFlight     map[string]*msgContext
 	inFlightCond *sync.Cond
 }
 
-func defInt(envVarName string, defValue int) int {
-	defStr := os.Getenv(envVarName)
-	if defStr == "" {
-		return defValue
-	}
-	parsedInt, err := strconv.ParseInt(defStr, 10, 32)
-	if err != nil {
-		log.Errorf("Invalid string in env var %s", envVarName)
-		return defValue
-	}
-	return int(parsedInt)
-}
-
-// CobraInit retruns a cobra command to configure this Kafka
+// CobraInit retruns a cobra command to configure this KafkaBridge
 func (k *KafkaBridge) CobraInit() (cmd *cobra.Command) {
 	cmd = &cobra.Command{
 		Use:   "kafka",
@@ -99,19 +56,7 @@ func (k *KafkaBridge) CobraInit() (cmd *cobra.Command) {
 			return
 		},
 		PreRunE: func(cmd *cobra.Command, args []string) (err error) {
-			if k.Conf.TopicOut == "" {
-				return fmt.Errorf("No output topic specified for bridge to send events to")
-			}
-			if k.Conf.TopicIn == "" {
-				return fmt.Errorf("No input topic specified for bridge to listen to")
-			}
-			if k.Conf.ConsumerGroup == "" {
-				return fmt.Errorf("No consumer group specified")
-			}
-			if err = kldutils.AllOrNoneReqd(cmd, "tls-clientcerts", "tls-clientkey"); err != nil {
-				return
-			}
-			if err = kldutils.AllOrNoneReqd(cmd, "sasl-username", "sasl-password"); err != nil {
+			if err = k.kafka.CobraPreRunE(cmd); err != nil {
 				return
 			}
 			if k.Conf.RPC.URL == "" {
@@ -123,80 +68,11 @@ func (k *KafkaBridge) CobraInit() (cmd *cobra.Command) {
 			return
 		},
 	}
-	defBrokerList := strings.Split(os.Getenv("KAFKA_BROKERS"), ",")
-	defTLSenabled, _ := strconv.ParseBool(os.Getenv("KAFKA_TLS_ENABLED"))
-	defTLSinsecure, _ := strconv.ParseBool(os.Getenv("KAFKA_TLS_INSECURE"))
-	cmd.Flags().StringArrayVarP(&k.Conf.Brokers, "brokers", "b", defBrokerList, "Comma-separated list of bootstrap brokers")
-	cmd.Flags().StringVarP(&k.Conf.ClientID, "clientid", "i", os.Getenv("KAFKA_CLIENT_ID"), "Client ID (or generated UUID)")
-	cmd.Flags().StringVarP(&k.Conf.ConsumerGroup, "consumer-group", "g", os.Getenv("KAFKA_CONSUMER_GROUP"), "Client ID (or generated UUID)")
+	k.kafka.CobraInit(cmd)
 	cmd.Flags().IntVarP(&k.Conf.MaxInFlight, "maxinflight", "m", defInt("KAFKA_MAX_INFLIGHT", 10), "Maximum messages to hold in-flight")
 	cmd.Flags().StringVarP(&k.Conf.RPC.URL, "rpc-url", "r", os.Getenv("ETH_RPC_URL"), "JSON/RPC URL for Ethereum node")
 	cmd.Flags().IntVarP(&k.Conf.MaxTXWaitTime, "tx-timeout", "x", defInt("ETH_TX_TIMEOUT", 300), "Maximum wait time for an individual transaction (seconds)")
-	cmd.Flags().StringVarP(&k.Conf.TopicIn, "topic-in", "t", os.Getenv("KAFKA_TOPIC_IN"), "Topic to listen to")
-	cmd.Flags().StringVarP(&k.Conf.TopicOut, "topic-out", "T", os.Getenv("KAFKA_TOPIC_OUT"), "Topic to send events to")
-	cmd.Flags().StringVarP(&k.Conf.TLS.ClientCertsFile, "tls-clientcerts", "c", os.Getenv("KAFKA_TLS_CLIENT_CERT"), "A client certificate file, for mutual TLS auth")
-	cmd.Flags().StringVarP(&k.Conf.TLS.PrivateKeyFile, "tls-clientkey", "k", os.Getenv("KAFKA_TLS_CLIENT_KEY"), "A client private key file, for mutual TLS auth")
-	cmd.Flags().StringVarP(&k.Conf.TLS.CACertsFile, "tls-cacerts", "C", os.Getenv("KAFKA_TLS_CA_CERTS"), "CA certificates file (or host CAs will be used)")
-	cmd.Flags().BoolVarP(&k.Conf.TLS.Enabled, "tls-enabled", "e", defTLSenabled, "Encrypt network connection with TLS (SSL)")
-	cmd.Flags().BoolVarP(&k.Conf.InsecureSkipVerify, "tls-insecure", "z", defTLSinsecure, "Disable verification of TLS certificate chain")
-	cmd.Flags().StringVarP(&k.Conf.SASL.Username, "sasl-username", "u", os.Getenv("KAFKA_SASL_USERNAME"), "Username for SASL authentication")
-	cmd.Flags().StringVarP(&k.Conf.SASL.Password, "sasl-password", "p", os.Getenv("KAFKA_SASL_PASSWORD"), "Password for SASL authentication")
 	return
-}
-
-func (k *KafkaBridge) createTLSConfiguration() (t *tls.Config, err error) {
-
-	mutualAuth := k.Conf.TLS.ClientCertsFile != "" && k.Conf.TLS.PrivateKeyFile != ""
-	log.Debugf("Kafka TLS Enabled=%t Insecure=%t MutualAuth=%t ClientCertsFile=%s PrivateKeyFile=%s CACertsFile=%s",
-		k.Conf.TLS.Enabled, k.Conf.InsecureSkipVerify, mutualAuth, k.Conf.TLS.ClientCertsFile, k.Conf.TLS.PrivateKeyFile, k.Conf.TLS.CACertsFile)
-	if !k.Conf.TLS.Enabled {
-		return
-	}
-
-	var clientCerts []tls.Certificate
-	if mutualAuth {
-		var cert tls.Certificate
-		if cert, err = tls.LoadX509KeyPair(k.Conf.TLS.ClientCertsFile, k.Conf.TLS.PrivateKeyFile); err != nil {
-			log.Errorf("Unable to load client key/certificate: %s", err)
-			return
-		}
-		clientCerts = append(clientCerts, cert)
-	}
-
-	var caCertPool *x509.CertPool
-	if k.Conf.TLS.CACertsFile != "" {
-		var caCert []byte
-		if caCert, err = ioutil.ReadFile(k.Conf.TLS.CACertsFile); err != nil {
-			log.Errorf("Unable to load CA certificates: %s", err)
-			return
-		}
-		caCertPool = x509.NewCertPool()
-		caCertPool.AppendCertsFromPEM(caCert)
-	}
-
-	t = &tls.Config{
-		Certificates:       clientCerts,
-		RootCAs:            caCertPool,
-		InsecureSkipVerify: k.Conf.InsecureSkipVerify,
-	}
-	return
-}
-
-type saramaLogger struct {
-}
-
-func (s saramaLogger) Print(v ...interface{}) {
-	v = append([]interface{}{"[sarama] "}, v...)
-	log.Debug(v...)
-}
-
-func (s saramaLogger) Printf(format string, v ...interface{}) {
-	log.Debugf("[sarama] "+format, v...)
-}
-
-func (s saramaLogger) Println(v ...interface{}) {
-	v = append([]interface{}{"[sarama] "}, v...)
-	log.Debug(v...)
 }
 
 // MsgContext is passed for each message that arrives at the bridge
@@ -235,7 +111,7 @@ type msgContext struct {
 func (k *KafkaBridge) addInflightMsg(msg *sarama.ConsumerMessage) (pCtx *msgContext, err error) {
 	ctx := msgContext{
 		timeReceived: time.Now(),
-		origMsg:      fmt.Sprintf("%s:%d:%d", k.Conf.TopicIn, msg.Partition, msg.Offset),
+		origMsg:      fmt.Sprintf("%s:%d:%d", k.kafka.Conf.TopicIn, msg.Partition, msg.Offset),
 		saramaMsg:    msg,
 		bridge:       k,
 	}
@@ -282,7 +158,7 @@ func (a ctxByOffset) Less(i, j int) bool {
 // Looks at the other in-flight messages for the same partition, and works out if
 // we can move the offset forwards.
 // * Caller holds the inFlightCond mutex *
-func (k *KafkaBridge) setInFlightComplete(ctx *msgContext) (err error) {
+func (k *KafkaBridge) setInFlightComplete(ctx *msgContext, consumer KafkaConsumer) (err error) {
 
 	// Build an offset sorted list of the inflight
 	ctx.complete = true
@@ -316,7 +192,7 @@ func (k *KafkaBridge) setInFlightComplete(ctx *msgContext) (err error) {
 		// Update the offset
 		highestOffset := readyToAck[len(readyToAck)-1].saramaMsg
 		log.Infof("Marking offset %d:%d", highestOffset.Offset, highestOffset.Partition)
-		k.consumer.MarkOffset(highestOffset, "")
+		consumer.MarkOffset(highestOffset, "")
 	}
 
 	return
@@ -352,8 +228,8 @@ func (c *msgContext) Reply(replyMessage kldmessages.ReplyWithHeaders) {
 	replyHeaders.Elapsed = c.replyTime.Sub(c.timeReceived).Seconds()
 	c.replyBytes, _ = json.Marshal(replyMessage)
 	log.Infof("Sending reply: %s", c)
-	c.bridge.producer.Input() <- &sarama.ProducerMessage{
-		Topic:    c.bridge.Conf.TopicOut,
+	c.bridge.kafka.producer.Input() <- &sarama.ProducerMessage{
+		Topic:    c.bridge.kafka.Conf.TopicOut,
 		Key:      sarama.StringEncoder(c.key),
 		Metadata: c.origMsg,
 		Value:    c,
@@ -385,127 +261,19 @@ func (c msgContext) Encode() ([]byte, error) {
 
 // NewKafkaBridge creates a new KafkaBridge
 func NewKafkaBridge() *KafkaBridge {
-	var kf saramaKafkaFactory
 	mp := newMsgProcessor()
-	k := KafkaBridge{
-		factory:      &kf,
+	k := &KafkaBridge{
 		processor:    mp,
 		inFlight:     make(map[string]*msgContext),
 		inFlightCond: sync.NewCond(&sync.Mutex{}),
 	}
-	return &k
+	k.kafka = NewKafkaCommon(k)
+	return k
 }
 
-func (k *KafkaBridge) connect() (err error) {
-
-	sarama.Logger = k.saramaLogger
-	clientConf := cluster.NewConfig()
-
-	var tlsConfig *tls.Config
-	if tlsConfig, err = k.createTLSConfiguration(); err != nil {
-		return
-	}
-
-	if k.Conf.SASL.Username != "" && k.Conf.SASL.Password != "" {
-		clientConf.Net.SASL.Enable = true
-		clientConf.Net.SASL.User = k.Conf.SASL.Username
-		clientConf.Net.SASL.Password = k.Conf.SASL.Password
-	}
-
-	clientConf.Producer.Return.Successes = true
-	clientConf.Producer.Return.Errors = true
-	clientConf.Producer.RequiredAcks = sarama.WaitForLocal
-	clientConf.Producer.Flush.Frequency = 500 * time.Millisecond
-	clientConf.Consumer.Return.Errors = true
-	clientConf.Group.Return.Notifications = true
-	clientConf.Net.TLS.Enable = (tlsConfig != nil)
-	clientConf.Net.TLS.Config = tlsConfig
-	clientConf.ClientID = k.Conf.ClientID
-	if clientConf.ClientID == "" {
-		clientConf.ClientID = kldutils.UUIDv4()
-	}
-	log.Debugf("Kafka ClientID: %s", clientConf.ClientID)
-
-	log.Debugf("Kafka Bootstrap brokers: %s", k.Conf.Brokers)
-	if k.client, err = k.factory.newClient(k, clientConf); err != nil {
-		log.Errorf("Failed to create Kafka client: %s", err)
-		return
-	}
-	var brokers []string
-	for _, broker := range k.client.Brokers() {
-		brokers = append(brokers, broker.Addr())
-	}
-	log.Infof("Kafka Connected: %s", brokers)
-
-	// Connect the client
-	if k.rpc, err = rpc.Dial(k.Conf.RPC.URL); err != nil {
-		err = fmt.Errorf("JSON/RPC connection to %s failed: %s", k.Conf.RPC.URL, err)
-		return
-	}
-	k.processor.Init(k.rpc, k.Conf.MaxTXWaitTime)
-	log.Debug("JSON/RPC connected. URL=", k.Conf.RPC.URL)
-
-	return
-}
-
-func (k *KafkaBridge) producerErrorLoop() {
-	for err := range k.producer.Errors() {
-		k.inFlightCond.L.Lock()
-		// If we fail to send a reply, this is significant. We have a request in flight
-		// and we have probably already sent the message.
-		// Currently we panic, on the basis that we will be restarted by Docker
-		// to drive retry logic. In the future we might consider recreating the
-		// producer and attempting to resend the message a number of times -
-		// keeping a retry counter on the msgContext object
-		origMsg := err.Msg.Metadata.(string)
-		ctx := k.inFlight[origMsg]
-		log.Errorf("Kafka producer failed for reply %s to origMsg %s: %s", ctx, origMsg, err)
-		panic(err)
-		// k.inFlightCond.L.Unlock() - unreachable while we have a panic
-	}
-	k.producerWG.Done()
-}
-
-func (k *KafkaBridge) producerSuccessLoop() {
-	for msg := range k.producer.Successes() {
-		k.inFlightCond.L.Lock()
-		origMsg := msg.Metadata.(string)
-		if ctx, ok := k.inFlight[origMsg]; ok {
-			log.Infof("Reply sent: %s", ctx)
-			// While still holding the lock, add this to the completed list
-			k.setInFlightComplete(ctx)
-			// We've reduced the in-flight count - wake any waiting consumer go func
-			k.inFlightCond.Broadcast()
-		} else {
-			// This should never happen. Represents a logic bug that must be diagnosed.
-			err := fmt.Errorf("Received confirmation for message not in in-flight map: %s", origMsg)
-			panic(err)
-		}
-		k.inFlightCond.L.Unlock()
-	}
-	k.producerWG.Done()
-}
-
-func (k *KafkaBridge) startProducer() (err error) {
-
-	log.Debugf("Kafka Producer Topic=%s", k.Conf.TopicOut)
-	if k.producer, err = k.client.newProducer(k); err != nil {
-		log.Errorf("Failed to create Kafka producer: %s", err)
-		return
-	}
-
-	k.producerWG.Add(2)
-
-	go k.producerErrorLoop()
-
-	go k.producerSuccessLoop()
-
-	log.Infof("Kafka Created producer")
-	return
-}
-
-func (k *KafkaBridge) consumerMessagesLoop() {
-	for msg := range k.consumer.Messages() {
+// ConsumerMessagesLoop - goroutine to process messages
+func (k *KafkaBridge) ConsumerMessagesLoop(consumer KafkaConsumer, producer KafkaProducer, wg *sync.WaitGroup) {
+	for msg := range consumer.Messages() {
 		k.inFlightCond.L.Lock()
 		log.Infof("Kafka consumer received message: Partition=%d Offset=%d", msg.Partition, msg.Offset)
 
@@ -528,61 +296,70 @@ func (k *KafkaBridge) consumerMessagesLoop() {
 			msgCtx.Reply(errMsg)
 		}
 	}
-	k.consumerWG.Done()
+	wg.Done()
 }
 
-func (k *KafkaBridge) startConsumer() (err error) {
+// ProducerErrorLoop - goroutine to process producer errors
+func (k *KafkaBridge) ProducerErrorLoop(consumer KafkaConsumer, producer KafkaProducer, wg *sync.WaitGroup) {
+	for err := range producer.Errors() {
+		k.inFlightCond.L.Lock()
+		// If we fail to send a reply, this is significant. We have a request in flight
+		// and we have probably already sent the message.
+		// Currently we panic, on the basis that we will be restarted by Docker
+		// to drive retry logic. In the future we might consider recreating the
+		// producer and attempting to resend the message a number of times -
+		// keeping a retry counter on the msgContext object
+		origMsg := err.Msg.Metadata.(string)
+		ctx := k.inFlight[origMsg]
+		log.Errorf("Kafka producer failed for reply %s to origMsg %s: %s", ctx, origMsg, err)
+		panic(err)
+		// k.inFlightCond.L.Unlock() - unreachable while we have a panic
+	}
+	wg.Done()
+}
 
-	log.Debugf("Kafka Consumer Topic=%s ConsumerGroup=%s", k.Conf.TopicIn, k.Conf.ConsumerGroup)
-	if k.consumer, err = k.client.newConsumer(k); err != nil {
-		log.Errorf("Failed to create Kafka consumer: %s", err)
+// ProducerSuccessLoop - goroutine to process producer successes
+func (k *KafkaBridge) ProducerSuccessLoop(consumer KafkaConsumer, producer KafkaProducer, wg *sync.WaitGroup) {
+	for msg := range producer.Successes() {
+		k.inFlightCond.L.Lock()
+		origMsg := msg.Metadata.(string)
+		if ctx, ok := k.inFlight[origMsg]; ok {
+			log.Infof("Reply sent: %s", ctx)
+			// While still holding the lock, add this to the completed list
+			k.setInFlightComplete(ctx, consumer)
+			// We've reduced the in-flight count - wake any waiting consumer go func
+			k.inFlightCond.Broadcast()
+		} else {
+			// This should never happen. Represents a logic bug that must be diagnosed.
+			err := fmt.Errorf("Received confirmation for message not in in-flight map: %s", origMsg)
+			panic(err)
+		}
+		k.inFlightCond.L.Unlock()
+	}
+	wg.Done()
+}
+
+func (k *KafkaBridge) connect() (err error) {
+	// Connect the client
+	if k.rpc, err = rpc.Dial(k.Conf.RPC.URL); err != nil {
+		err = fmt.Errorf("JSON/RPC connection to %s failed: %s", k.Conf.RPC.URL, err)
 		return
 	}
+	k.processor.Init(k.rpc, k.Conf.MaxTXWaitTime)
+	log.Debug("JSON/RPC connected. URL=", k.Conf.RPC.URL)
 
-	k.consumerWG.Add(3)
-	go func() {
-		for err := range k.consumer.Errors() {
-			log.Error("Kafka consumer failed:", err)
-		}
-		k.consumerWG.Done()
-	}()
-	go func() {
-		for ntf := range k.consumer.Notifications() {
-			log.Debugf("Kafka consumer rebalanced. Current=%+v", ntf.Current)
-		}
-		k.consumerWG.Done()
-	}()
-	go k.consumerMessagesLoop()
-
-	log.Infof("Kafka Created consumer")
 	return
 }
 
 // Start kicks off the bridge
 func (k *KafkaBridge) Start() (err error) {
 
+	// Connect the RPC URL
 	if err = k.connect(); err != nil {
 		return
 	}
-	if err = k.startConsumer(); err != nil {
-		return
-	}
-	if err = k.startProducer(); err != nil {
-		return
-	}
 
-	k.signals = make(chan os.Signal, 1)
-	signal.Notify(k.signals, os.Interrupt)
-	for {
-		select {
-		case <-k.signals:
-			k.producer.AsyncClose()
-			k.consumer.Close()
-			k.producerWG.Wait()
-			k.consumerWG.Wait()
-
-			log.Infof("Kafka Bridge complete")
-			return
-		}
-	}
+	// Defer to KafkaCommon processing
+	err = k.kafka.Start()
+	return
 }
